Moved CouponListAdapter.java
